test(unicode): move unicode & tests to its own module
