+++ an-rope/src/test/mod.rs
@@ -16,9 +16,6 @@ impl Arbitrary for Rope {
 
 }
 
-#[cfg(test)]
-mod unicode;
-
 #[test]
 fn delete_test_1() {
     let mut r = Rope::from("this is not fine".to_string());
@@ -421,25 +418,25 @@ mod properties {
     //     quickcheck(prop as fn(String, usize) -> TestResult);
     // }
 
-    // #[test]
-    // fn rope_insert_char_is_string_insert_char() {
-    //     fn prop(a: String, ch: char, i: usize) -> TestResult {
-    //         // if the index is greater than the string's length...
-    //         if i > a.len() {
-    //             // ..skip the test
-    //             return TestResult::discard()
-    //         }
-    //
-    //         let mut rope = Rope::from(a.clone());
-    //         rope.insert(i, ch);
-    //
-    //         let mut string = a;
-    //         string.insert(i, ch);
-    //
-    //         TestResult::from_bool(rope == string)
-    //     }
-    //     quickcheck(prop as fn(String, char, usize) -> TestResult);
-    // }
+    #[test]
+    fn rope_insert_char_is_string_insert_char() {
+        fn prop(a: String, ch: char, i: usize) -> TestResult {
+            // if the index is greater than the string's length...
+            if i > a.len() {
+                // ..skip the test
+                return TestResult::discard()
+            }
+
+            let mut rope = Rope::from(a.clone());
+            rope.insert(i, ch);
+
+            let mut string = a;
+            string.insert(i, ch);
+
+            TestResult::from_bool(rope == string)
+        }
+        quickcheck(prop as fn(String, char, usize) -> TestResult);
+    }
 
     #[cfg(feature = "unstable")]
     #[test]
